Reorganize 'downloaders' exceptions
Since most exceptions are bare ancestors of python's builtin Exception class now, it feels just too noisy to have a single module for each of them. Therefore they are all merged into a single 'exceptions' module within the 'downloaders' package.
