--- package_control/downloaders/binary_not_found_error.py
@@ -1,6 +1,0 @@
-
-class BinaryNotFoundError(Exception):
-
-    """Necessary executable is not found in the PATH on the system."""
-
-    pass
