--- package_control/downloaders/downloader_exception.py
@@ -1,6 +1,0 @@
-
-class DownloaderException(Exception):
-
-    """Downloader can not download a URL."""
-
-    pass
