--- package_control/downloaders/non_http_error.py
@@ -1,4 +1,0 @@
-
-class NonHttpError(Exception):
-    """A downloader had a non-clean exit, but not due to an HTTP error."""
-    pass
