+++ package_control/downloaders/wget_downloader.py
@@ -7,10 +7,10 @@ from ..console_write import console_write
 from .caching_downloader import CachingDownloader
 from .cli_downloader import CliDownloader
 from .decoding_downloader import DecodingDownloader
-from .downloader_exception import DownloaderException
+from .exceptions import DownloaderException
+from .exceptions import NonCleanExitError
+from .exceptions import NonHttpError
 from .limiting_downloader import LimitingDownloader
-from .non_clean_exit_error import NonCleanExitError
-from .non_http_error import NonHttpError
 
 
 class WgetDownloader(CliDownloader, DecodingDownloader, LimitingDownloader, CachingDownloader):
