--- package_control/downloaders/win_downloader_exception.py
@@ -1,11 +1,0 @@
-from .downloader_exception import DownloaderException
-
-
-class WinDownloaderException(DownloaderException):
-
-    """
-    If the WinInetDownloader ran into a windows-specific error. The means we
-    should retry with the UrllibDownloader.
-    """
-
-    pass
